requirements.txt
streamlit
openai
python-dotenv
streamlit-webrtc
soundfile
av
audio-recorder-streamlit
streamlit-mic-recorder
pyperclip